fixes input func (str) calls

+++ include/weapon_stats.h
@@ -11,5 +11,5 @@ extern std::string bows[];
 int weapon_name();
 std::string level_chk(int);
 float substat_chk(int, std::string);
-float passive_chk(float);
+float passive_chk(int);
 #endif
